+++ src/GafferSceneModule/QueryBinding.cpp
@@ -47,6 +47,7 @@
 #include "GafferScene/ExistenceQuery.h"
 #include "GafferScene/FilterQuery.h"
 #include "GafferScene/OptionQuery.h"
+#include "GafferScene/PointInstancerQuery.h"
 #include "GafferScene/PrimitiveQuery.h"
 #include "GafferScene/PrimitiveVariableQuery.h"
 #include "GafferScene/SceneStats.h"
@@ -292,6 +293,7 @@ void GafferSceneModule::bindQueries()
 		;
 	}
 
+	GafferBindings::DependencyNodeClass< GafferScene::PointInstancerQuery >();
 	GafferBindings::DependencyNodeClass< GafferScene::PrimitiveQuery >();
 	GafferBindings::DependencyNodeClass< GafferScene::ExistenceQuery >();
 	GafferBindings::DependencyNodeClass< GafferScene::FilterQuery >();